1. Add metering type to inventory management 2. Fixed problems in updating customer selection list 3. Fixed problems in hardware configuration

+++ yukon-client/stars/src/com/cannontech/database/data/stars/hardware/InventoryBase.java
@@ -34,11 +34,25 @@ public class InventoryBase extends DBPersistent {
     }
     
     /**
-     * This method is to be called from the delete() method of all subclasses of InventoryBase (e.g. com.cannontech.database.data.stars.hardware.LMHardwareBase)
+     * When deleteInv = false, this method is used to remove inventory from a customer account.
+     * This method should not be used alone with deleteInv = true, instead it will be invoked
+     * from the delete() method of this class and all its subclasses with deleteInv set to true
+     * (e.g. com.cannontech.database.data.stars.hardware.LMHardwareBase).
      */
-    public void deleteInventoryBase() throws java.sql.SQLException {
-        delete( "ECToInventoryMapping", "InventoryID", getInventoryBase().getInventoryID() );
-        getInventoryBase().delete();
+    public void deleteInventoryBase(boolean deleteInv) throws java.sql.SQLException {
+    	if (deleteInv) {
+			// delete from LMHardwareEvent
+			com.cannontech.database.data.stars.event.LMHardwareEvent.deleteAllLMHardwareEvents(
+					getInventoryBase().getInventoryID(), getDbConnection() );
+			
+			delete( "ECToInventoryMapping", "InventoryID", getInventoryBase().getInventoryID() );
+			getInventoryBase().delete();
+    	}
+    	else {
+			getInventoryBase().setAccountID( new Integer(com.cannontech.common.util.CtiUtilities.NONE_ID) );
+			getInventoryBase().setRemoveDate( new java.util.Date() );
+			getInventoryBase().update();
+    	}
     }
 
     public void delete() throws java.sql.SQLException {
@@ -48,7 +62,7 @@ public class InventoryBase extends DBPersistent {
     	hw.setDbConnection( getDbConnection() );
     	hw.deleteLMHardwareBase( true );
     	
-    	deleteInventoryBase();
+    	deleteInventoryBase( true );
     }
 
     public void add() throws java.sql.SQLException {
